Show and calculate the transfered bytes in the Bandwith Graph. Removed compiler warnings.
git-svn-id: http://svn.code.sf.net/p/retroshare/code/trunk@3714 b45a01b8-16f6-495d-af2f-9b41ad6348cc

+++ retroshare-gui/src/gui/bwgraph/bwgraph.h
@@ -55,7 +55,7 @@ protected:
   
 private slots:
   /** Adds new data to the graph */
-  void updateGraph(quint64 bytesRead, quint64 bytesWritten);
+  void updateGraph(qreal bytesRead, qreal bytesWritten);
   /** Called when settings button is toggled */
   void showSettingsFrame(bool show);
   /** Called when the settings button is toggled */
